move logout and hide completed to settings-view and add navigation
