+++ src/views/TasksView.vue
@@ -1,26 +1,17 @@
 <script setup>
-import { ref, onMounted } from 'vue'
+import { onMounted } from 'vue'
 import { useTasksStore } from '@/stores/tasksStore'
 import CreateTask from '@/components/CreateTask.vue'
 import TaskCard from '@/components/TaskCard.vue'
 import { useUserStore } from '@/stores/userStore'
-import { useRouter } from 'vue-router'
 import SortTasks from '@/components/SortTasks.vue'
 import { storeToRefs } from 'pinia'
 
-const router = useRouter()
 const tasksStore = useTasksStore()
 const userStore = useUserStore()
 
 const { taskSortered } = storeToRefs(tasksStore)
-const { user, hideCompletedSetting } = storeToRefs(userStore)
-
-const newHideCompletedSetting = ref(hideCompletedSetting.value)
-
-const logOut = async () => {
-  await userStore.signOut()
-  router.push('/login')
-}
+const { user } = storeToRefs(userStore)
 
 onMounted(async () => {
   await userStore.fetchUserHideCompletedSetting()
@@ -50,16 +41,7 @@ onMounted(async () => {
       <p v-else-if="!taskSortered">Loading tasks...</p>
       <p v-else>No tasks available.</p>
     </template>
-    <button @click="logOut()" class="mt-4 bg-red-500 text-white px-4 py-2 rounded">Logout</button>
     <button><router-link to="/settings" class="mt-4 bg-blue-500 text-white px-4 py-2 rounded">Settings</router-link></button>
-    <div class="flex flex-col">
-      <div class="form-control w-52">
-        <label class="cursor-pointer label">
-          <span class="label-text">Hide completed tasks</span> 
-          <input type="checkbox" class="toggle toggle-accent" v-model="newHideCompletedSetting" @change="userStore.updateUserHideCompletedSetting(newHideCompletedSetting)" />
-        </label>
-      </div>
-    </div>
   </main>
 </template>
